requirements.txt
celery==5.4.0
cron-descriptor==1.4.5
dj-rest-auth==6.0.0
Django==5.0.7
django-allauth==0.63.6
django-celery-beat==2.7.0
django-cors-headers==4.4.0
django-phonenumber-field==8.0.0
django-phonenumbers==1.0.1
django-redis==5.4.0
django-rest-passwordreset==1.4.1
django-rest-swagger==2.2.0
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
environ==1.0
pathspec==0.12.1
phonenumbers==8.13.40
platformdirs==4.2.2
psycopg2-binary==2.9.9
PyJWT==2.8.0
redis==5.0.8
rest-framework-simplejwt==0.0.2
django-cron==0.6.0
django-crontab==0.7.1
